fix: change np.int and np.bool to int and bool; no longer supported

+++ pychunkedgraph/app/meshing/common.py
@@ -72,7 +72,7 @@ def handle_get_manifest(table_id, node_id):
     if "bounds" in request.args:
         bounds = request.args["bounds"]
         bounding_box = np.array(
-            [b.split("-") for b in bounds.split("_")], dtype=np.int
+            [b.split("-") for b in bounds.split("_")], dtype=int
         ).T
     else:
         bounding_box = None
